Made 'GameManager' Script
Created a 'GameManager' script that will be used to transfer values
across scenes. For now, it can transfer values for the CountdownTimer
from one scene to another.

NOTE: The first scene the GameManager loads that has a 'CountdownTimer'
prefab should override the values set on the prefab with the values that
are stored in the GameManager. These values are private, to make sure
that nothing other then the GameManager can mess with them, so if they
need to be changed, change it in the script.