Basic Menu and Teleportation
Now there is a (very) basic main menu which can start or quit the game. We haven't arraged the build orders yet so it does not really start the game.

Also there are now teleportation pads. It uses the NavMesh system to do teleportation, but works with the WASD controls. Step on the first pad and it takes you to the second. Get off second pad and step back on to arrive at the first again.

There is a new scene, a new prefab, a new material, anda few new scripts

-Changed tag of Player to "Player". This is necessary for the Portals to work